refactor: extract SpotlightModal component
